update style and formating
