+++ src/ScottLogic.Internal.Training/ScottLogic.Internal.Training.Matcher/OrderMatcher.cs
@@ -10,7 +10,6 @@ namespace ScottLogic.Internal.Training.Matcher
 
         public OrderMatcher()
         {
-            CurrentTrade = null;
         }
 
         public bool ProcessOrder(Order currentOrder)
